+++ assets/css/styles/supervision-tree.css
@@ -1,3 +1,7 @@
+/*
+* These styles are meant to be used in SupervisionTree JS hook
+*/
+
 [data-cy-container] {
   --grid: rgba(148, 163, 184, 0.12);
   background:
