+++ assets/js/hooks/supervision_tree/constants.js
@@ -1,0 +1,13 @@
+export const LAYOUT_DEBOUNCE_MS = 50;
+export const OVERLAY_DEBOUNCE_MS = 80;
+export const OVERLAY_MIN_ZOOM = 0.45;
+export const FADE_MS = 200;
+
+export const TOPOLOGY_FIELDS = new Set([
+  'name',
+  'type',
+  'has_children?',
+  'child_count',
+  'children_keys',
+  'parent_key',
+]);
